Create a single module for trackbars
